Add basic PID loop

+++ main/camera.h
@@ -5,8 +5,8 @@
 // Class to read and interpret data from the HuskyLens camera
 class Camera {
 public:
-    int angle;
-    int offset;
+    float angle;
+    float offset;
 
     void init();
